% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_annotation}
\alias{add_annotation}
\title{Add Annotation}
\usage{
add_annotation(res, annotation_file)
}
\arguments{
\item{res}{Results table obtained from MAGESTICA (MAGESTICA_output$results).}

\item{annotation_file}{Annotation file with rownames corresponding to the barcodes.}
}
\value{
Results table with annotation information.
}
\description{
Given the results table from MAGESTICA and an annotation file with rownames corresponding to the barcodes, returns the results table with annotation information.
}
