% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_monotonicity.R
\name{add_monotonicity}
\alias{add_monotonicity}
\title{Add monotonicity filter}
\usage{
add_monotonicity(df)
}
\arguments{
\item{df}{Data frame obtained with function df_consecutive_LFCs.}
}
\value{
Data frame with a statistic addded (dir) on which we can filter for monotonicity.
}
\description{
Adds statistic which we refer to as LFC direction (dir) to the data frame obtained with df_consecutive_LFCs. We can filter on this statistic for monotonicity.
}
