% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_consecutive_LFCs.R
\name{df_consecutive_LFCs}
\alias{df_consecutive_LFCs}
\title{Create Table of Trajectory of LFCs}
\usage{
df_consecutive_LFCs(dds)
}
\arguments{
\item{dds}{DESeqDataSet corresponding to MAGESTICA pipeline (MAGESTICA_output$dds.MAGESTICA).}
}
\value{
Table containing the LFCs over consecutive timepoints (T1vsT2, T2vsT3, ...) for every barcode.
}
\description{
Given a DESeqDataSet from the output of MAGESTICA, return a table containing the LFCs over consecutive timepoints (T1vsT2, T2vsT3, ...) for every barcode.
}
\examples{
data(MAGESTICA_output)
df <- df_consecutive_LFCs(MAGESTICA_output$dds.MAGESTICA)
}
