% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_control.R
\name{get_control}
\alias{get_control}
\title{Get Subset corresponding to Control Barcodes}
\usage{
get_control(df, control)
}
\arguments{
\item{df}{A data frame with barcodes as the rownames e.g. count matrix or results table.}

\item{control}{A vector containing the names of the control barcodes.}
}
\value{
A subset of the data frame containing only the control features.
}
\description{
Given a list of the control barcodes, returns the control barcodes in the data frame.
}
\examples{
data(counts_MAGESTICA, control_bcs)
counts_MAGESTICA_control <- get_control(counts_MAGESTICA, control_bcs)
}
