% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indices_control.R
\name{get_indices_control}
\alias{get_indices_control}
\title{Get Indices corresponding to the Control Barcodes (Helper Function)}
\usage{
get_indices_control(df, control)
}
\arguments{
\item{df}{A data frame with the barcodes as rownames e.g. count matrix or results table.}

\item{control}{A vector containing the control barcodes.}
}
\value{
A vector containing the indices corresponding to the control barcodes in given data frame.
}
\description{
Given a list of the control barcodes, returns the indices of the control barcodes in the data frame.
}
