% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_plot.R
\name{MA_plot}
\alias{MA_plot}
\title{MA-Plot}
\usage{
MA_plot(res, alpha = 0.1, thresholds = c(0, 0))
}
\arguments{
\item{res}{Results table obtained from MAGESTICA (MAGESTICA_output$results).}

\item{alpha}{Significance cutoff (by default 0.1). If the adjusted p-value cutoff (FDR) used was a value other than 0.1, alpha should be set to that value.}

\item{thresholds}{Thresholds against which testing was done, calculated in MAGESTICA (MAGESTICA_output$min.fitness.thresholds).}
}
\value{
An MA-plot indicating thresholds against which was tested.
}
\description{
Generates an MA-plot from the results of MAGESTICA. Thresholds indicate the minimum fitness effect against which was tested.
}
\examples{
data(MAGESTIC_output, control_bcs)
MA_plot(MAGESTICA_output$results, thresholds = MAGESTICA_output$min.fitness.thresholds)
}
