% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGESTICA.R
\name{MAGESTICA}
\alias{MAGESTICA}
\title{MAGESTIC analysis}
\usage{
MAGESTICA(
  counts,
  coldata,
  control = NULL,
  alpha = 0.1,
  quantile_range = c(0.1, 0.9),
  norm_method = "GMPR",
  pseudo_count = 0
)
}
\arguments{
\item{counts}{A raw count matrix with the rows corresponding to the barcodes and the columns to the samples.}

\item{coldata}{A table with the rows corresponding to the samples and the columns corresponding to the metadata. Columns must contain at least timepoint and replicate.}

\item{control}{A vector containing the names of the control barcodes.}

\item{alpha}{Significance cutoff (by default 0.1). If the adjusted p-value cutoff (FDR) used was a value other than 0.1, alpha should be set to that value.}

\item{quantile_range}{Vector of two values, relating to the quantile points which are used to calculate thresholds. (from the WT LFCs) to see which .}

\item{norm_method}{Which normalization approach to employ. Two normalization approaches are available: Geometric Mean of Pairwise Ratios (GMPR) and Relative Log Expression (RLE). Default is GMPR.}

\item{pseudo_count}{A positive numeric value for the pseudo-counts to be added to account for the influence of zeros during analysis. Default is 1.}
}
\value{
List with output of MAGESTICA. It contains: The normalized counts matrix (norm.counts), DESeqDataSet used for analysis (dds.MAGESTICA), results table of the analysis (results), size factors used during normalization (size.factors), LFC against which is tested for the analysis.
}
\description{
The MAGESTIC pipeline. Function needs at least a count matrix, a meta file and list of the control barcode names to work.
}
\examples{
data(counts_MAGESTIC, metadata_MAGESTIC, control_bcs)
MAGESTIC_output <- MAGESTIC(counts_MAGESTIC, metadata_MAGESTIC, control_bcs)
MAGESTIC_output <- MAGESTIC(counts = counts_MAGESTIC, coldata = metadata_MAGESTIC, control = control_bcs, alpha = 0.1, quantile_range = c(0.025, 0.975), norm_method = "RLE", pseudo_counts = 0)
}
