% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_counts.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot Count Trajectories}
\usage{
plot_counts(counts, meta, bc)
}
\arguments{
\item{counts}{A count matrix (raw or normalized) with the rows corresponding to the barcodes and the columns to the samples.}

\item{meta}{A table with the rows corresponding to the samples and the columns corresponding to the metadata. Columns must contain at least timepoint and replicate.}

\item{bc}{A barcode name}
}
\value{
A plot showing the trajectories of the different replicates over time for the given barcode.
}
\description{
Given a count matrix (raw or normalized), metadata and barcode names, returns a plot showing trajectories of the different replicates over time for the given barcode.
}
\examples{
data(MAGESTICA_output, metadata_MAGESTIC, control_bcs)
bc <- rownames(MAGESTICA_output$results)[1] # First barcode of dataframe (random example)
plot_counts(MAGESTICA_output$norm.counts, metadata_MAGESTIC, bc = bc)
}
