% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitness_distr.R
\name{plot_fitness_distr}
\alias{plot_fitness_distr}
\title{Plot Fitness Distribution}
\usage{
plot_fitness_distr(res, control, thresholds = c(0, 0))
}
\arguments{
\item{res}{Results table obtained from MAGESTICA (MAGESTICA_output$results).}

\item{control}{Vector containing the control barcode names.}

\item{thresholds}{Thresholds against which testing was done, calculated in MAGESTICA (MAGESTICA_output$min.fitness.thresholds).}
}
\value{
Plot of fitness distribution with thresholds against which were tested.
}
\description{
Generates a plot showing fitness (= LFC) distributions of both edited and WT group.
}
\examples{
data(MAGESTICA_output, control_bcs)
plot_fitness_distr(MAGESTICA_output$results, control = control_bcs, thresholds = MAGESTICA_output$min.fitness.thresholds)
}
