% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_consecutive_timepoints_LFCshrinkage.R
\name{results_consecutive_timepoints_LFCshrinkage}
\alias{results_consecutive_timepoints_LFCshrinkage}
\title{Helper function}
\usage{
results_consecutive_timepoints_LFCshrinkage(dds)
}
\description{
DO NOT USE
}
