% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/results_first_vs_other_timepoints_LFCshrinkage.R
\name{results_first_vs_other_timepoints_LFCshrinkage}
\alias{results_first_vs_other_timepoints_LFCshrinkage}
\title{Helper function}
\usage{
results_first_vs_other_timepoints_LFCshrinkage(dds)
}
\description{
DO NOT USE
}
