% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot}
\usage{
volcano_plot(res, alpha = 0.1, thresholds = c(0, 0))
}
\arguments{
\item{res}{Results table obtained from MAGESTICA (MAGESTICA_output$results).}

\item{alpha}{Significance cutoff (by default 0.1). If the adjusted p-value cutoff (FDR) used was a value other than 0.1, alpha should be set to that value.}

\item{thresholds}{Thresholds against which testing was done, calculated in MAGESTICA (MAGESTICA_output$min.fitness.thresholds).}
}
\value{
A volcano plot indicating thresholds against which were tested.
}
\description{
Generates a volcano plot from the results of MAGESTICA. Thresholds indicate the minimum fitness effect against which was tested.
}
\examples{
data(MAGESTIC_output, control_bcs)
volcano_plot(MAGESTICA_output$results, thresholds = MAGESTICA_output$min.fitness.thresholds)
}
